+++ src/main/java/com/example/demoAT/service/impl/CategoryServiceImpl.java
@@ -14,14 +14,13 @@ public class CategoryServiceImpl implements CategoryService {
     @Autowired
     private CategoryRepository categoryRepository;
 
-
     @Override
     public List<Category> getAllCategory() {
         return categoryRepository.findAll();
     }
 
     @Override
-    public Category getCategoryById(Integer id) {
+    public Category getCategoryById(Long id) {
         return categoryRepository.findById(id).orElse(null);
     }
 
@@ -31,12 +30,12 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void deleteCategory(Integer id) {
+    public void deleteCategory(Long id) {
         categoryRepository.deleteById(id);
     }
 
     @Override
-    public Category updateCategory(Integer id, Category category) {
+    public Category updateCategory(Long id, Category category) {
         Category exitstingCategory = categoryRepository.findById(id).orElse(null);
         if(exitstingCategory != null) {
             exitstingCategory.setCateName(category.getCateName());
